+++ src/commonMain/kotlin/com/metrolist/innertubex/cipher/QuickJsEngine.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.withContext
  */
 internal class QuickJsEngine {
     companion object {
-        private const val EVALUATION_TIMEOUT_MS = 10_000L
+        private const val EVALUATION_TIMEOUT_MS = 30_000L
 
         // Current ~2.6 MiB player scripts need more than 128 MiB while EJS builds their AST.
         private const val NATIVE_MEMORY_LIMIT_BYTES = 192L * 1024L * 1024L
@@ -51,7 +51,8 @@ internal class QuickJsEngine {
 
                         else -> {
                             if (c.code < 0x20) {
-                                append("\\u%04x".format(c.code))
+                                val hex = c.code.toString(16).padStart(4, '0')
+                                append("\\u$hex")
                             } else {
                                 append(c)
                             }